requirements.txt
pandas>=2.0
numpy>=1.24
scipy>=1.11
pyyaml>=6.0
jinja2>=3.1
